First try and ready!?!?
Download a kmz extract from the wandrer.earth site.  Convert this with gpsbabel to osm format. With a good text editor replace the - signs in the node etc lines to 18000000000 (to avoid colliding with id numbers in use on osm). And finally use osmconvert to convert the osm file to osm.pbf format (osmconvert wandrer.osm -o=wandrer.osm.pbf.
Move this file to the maps folder and generate new maps